Ignore the Python managed environment error
We've run into https://peps.python.org/pep-0668/, because we install a
bunch of Python packages via apt, and then try to also install the
awscli package via pip, which is not happy about having two sources of
packages.

We've always been doing this, so presumably it's safe to continue
doing so.