Build a Ruby 3.2.3/bookworm image (#21)
* Build a Ruby 3.2.3/bookworm image

The base image in question is

https://hub.docker.com/layers/library/ruby/3.2.3-bookworm/images/sha256-5590025acebb13cacd5a5a4e5b0733c18d841318b8b788996bb88ce1fc64c565

for buildpack, and

https://hub.docker.com/layers/library/ruby/3.2.3-slim-bookworm/images/sha256-7bd0053d5820b233c060775c8fe21d25a9f3ee1ae4fb04bc8fe7887f9c44a2f3

for the base image.

This should get us onto OpenSSL 3, which is the point of this, as well
as pick up a slightly newer Ruby version.

* Ignore the Python managed environment error

We've run into https://peps.python.org/pep-0668/, because we install a
bunch of Python packages via apt, and then try to also install the
awscli package via pip, which is not happy about having two sources of
packages.

We've always been doing this, so presumably it's safe to continue
doing so.